+++ CHANGELOG.md
@@ -5,7 +5,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `ConvertTo-DifferenceString`
+  - Improve how control characters (ASCII values 0-31 and 127) are converted
+    to their corresponding Unicode representations in the Control Pictures
+    block to output printable versions.
+
 ## [0.3.0] - 2024-09-02
+
+### Added
 
 - Public commands:
   - `Invoke-PesterJob`